Refactor tests, update reference image tests
The previous reference image tests did not send a file,
which is now required by the API. This adds the file to
the test payloads so that they work correctly.

Also refactors the tests into separate files under the
tests/ directory, as has been done for models/, views/,
and serializers/.
