+++ src/django/api/tests/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa: F401
+from .base import BoundarySyncAPITestCase
+from .reference_image import ReferenceImageViewTests
